More WIP (doesn't build)
